pyproject.toml
[project]
name = "wetzlar-karteikartenerkennung"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "google-cloud-vision>=3.11.0",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
]

[project.scripts]
main = "main:main"
